+++ contracts/exchange/ExchangeAdaptor.sol
@@ -28,6 +28,4 @@ contract ExchangeAdaptor {
     // EVENTS
 
     event OrderUpdated(uint id);
-    event Trade(address indexed seller, uint sell_how_much, address indexed sell_which_token,
-        address indexed buyer, uint buy_how_much, address indexed buy_which_token);
 }